+++ app/src/app/components/contents/downloads/FileGroupDownloadButton.tsx
@@ -1,11 +1,11 @@
 import { useParams } from "react-router-dom";
-import { download, getZipUrl } from "../../../../../lib/download";
-import { Button, buttonVariants } from "../../../Base/Button";
+import { download, getZipUrl } from "../../../../lib/download";
+import { Button, buttonVariants } from "../../Base/Button";
 import { FolderDown, Loader2 } from "lucide-react";
-import { usePacketOutletContext } from "../../../main/PacketOutlet";
-import { FileMetadata } from "../../../../../types";
+import { usePacketOutletContext } from "../../main/PacketOutlet";
+import { FileMetadata } from "../../../../types";
 import { useState } from "react";
-import { filesToSize } from "../../../../../helpers";
+import { filesToSize } from "../../../../helpers";
 import type { VariantProps } from "class-variance-authority";
 
 interface FileGroupDownloadButtonProps {
